+++ ci4/app/Interfaces/DeploymentAnnotation.php
@@ -1,0 +1,8 @@
+<?php namespace App\Interfaces;
+/**
+ * @property string $name
+ * @property string $value
+ */
+interface DeploymentAnnotation {
+
+}