+++ ci4/app/Interfaces/DeploymentAnnotationList.php
@@ -1,0 +1,7 @@
+<?php namespace App\Interfaces;
+/**
+ * @property DeploymentAnnotation[] $values
+ */
+interface DeploymentAnnotationList {
+
+}