+++ vue/src/constants.ts
@@ -205,6 +205,14 @@ import type {
 import type {
     DeploymentSpecificationUpdateRoleRuleDialog_Input
 } from "@/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateRoleRuleDialog.vue";
+import type DeploymentSpecificationUpdateDeploymentAnnotationsDialog
+    from "@/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateDeploymentAnnotationsDialog.vue";
+import type DeploymentSpecificationUpdateDeploymentAnnotationDialog
+    from "@/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateDeploymentAnnotationDialog.vue";
+import type DeploymentSpecificationUpdateIngressAnnotationsDialog
+    from "@/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateIngressAnnotationsDialog.vue";
+import type DeploymentSpecificationUpdateIngressAnnotationDialog
+    from "@/components/Dialogs/Setup/DeploymentSpecifications/UpdateDialogs/DeploymentSpecificationUpdateIngressAnnotationDialog.vue";
 
 export type Events = {
     confirm: ConfirmationDialog_Input;
@@ -293,6 +301,10 @@ export type Events = {
     deploymentSpecificationUpdateRoleRule: DeploymentSpecificationUpdateRoleRuleDialog_Input;
     deploymentSpecificationUpdateServiceAnnotations: DeploymentSpecificationUpdateServiceAnnotationsDialog_Input;
     deploymentSpecificationUpdateServiceAnnotation: DeploymentSpecificationUpdateServiceAnnotationDialog_Input;
+    deploymentSpecificationUpdateDeploymentAnnotations: DeploymentSpecificationUpdateDeploymentAnnotationsDialog;
+    deploymentSpecificationUpdateDeploymentAnnotation: DeploymentSpecificationUpdateDeploymentAnnotationDialog;
+    deploymentSpecificationUpdateIngressAnnotations: DeploymentSpecificationUpdateIngressAnnotationsDialog;
+    deploymentSpecificationUpdateIngressAnnotation: DeploymentSpecificationUpdateIngressAnnotationDialog;
     deploymentSpecificationUpdateInitContainers: DeploymentSpecificationUpdateInitContainersDialog_Input;
     deploymentSpecificationUpdatePostUpdateActions: DeploymentSpecificationUpdatePostUpdateActionsDialog_Input;
     deploymentSpecificationUpdateLabels: DeploymentSpecificationUpdateLabelsDialog_Input;