+++ vue/src/core/services/Deploy/models/DeploymentSpecificationDeploymentAnnotation.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 13-12-2024.
+ * Time: 17:06.
+ */
+import {DeploymentSpecificationDeploymentAnnotationDefinition} from './definitions/DeploymentSpecificationDeploymentAnnotationDefinition';
+
+export class DeploymentSpecificationDeploymentAnnotation extends DeploymentSpecificationDeploymentAnnotationDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}