+++ vue/src/core/services/Deploy/models/DeploymentSpecificationIngressAnnotation.ts
@@ -1,0 +1,14 @@
+/**
+ * Created by ModelParser
+ * Date: 13-12-2024.
+ * Time: 17:06.
+ */
+import {DeploymentSpecificationIngressAnnotationDefinition} from './definitions/DeploymentSpecificationIngressAnnotationDefinition';
+
+export class DeploymentSpecificationIngressAnnotation extends DeploymentSpecificationIngressAnnotationDefinition {
+
+    constructor(json?: any) {
+        super(json);
+    }
+
+}